+++ pkg/llm-d-inference-sim/text_completion.go
@@ -18,22 +18,16 @@ package llmdinferencesim
 
 import (
 	"encoding/json"
+	"strings"
 	"sync"
+	"time"
 
 	"github.com/llm-d/llm-d-inference-sim/pkg/common"
 	openaiserverapi "github.com/llm-d/llm-d-inference-sim/pkg/openai-server-api"
 	"github.com/valyala/fasthttp"
 )
 
-type textCompletionReqCtx struct {
-	baseRequestContext
-	req *textCompletionRequest
-}
-
-func (t *textCompletionReqCtx) request() request {
-	return t.req
-}
-
+// Implementation of request for /completions requests
 type textCompletionRequest struct {
 	openaiserverapi.TextCompletionRequest
 }
@@ -75,3 +69,112 @@ func (t *textCompletionRequest) createResponseContext(displayModel string, respo
 }
 
 var _ request = (*textCompletionRequest)(nil)
+
+// Implementation of requestContext for /completions requests
+type textCompletionReqCtx struct {
+	baseRequestContext
+	req *textCompletionRequest
+}
+
+func (t *textCompletionReqCtx) request() request {
+	return t.req
+}
+
+func (t *textCompletionReqCtx) kvCacheOnRequestStart() *openaiserverapi.Error {
+	if t.sim.config.EnableKVCache {
+		if err := t.sim.kvcacheHelper.OnRequestStart(t.request()); err != nil {
+			serverError := openaiserverapi.NewError(err.Error(), fasthttp.StatusInternalServerError, nil)
+			return &serverError
+		}
+	}
+	return nil
+}
+
+func (t *textCompletionReqCtx) kvCacheOnRequestEnd() {
+	if t.sim.config.EnableKVCache {
+		if err := t.sim.kvcacheHelper.OnRequestEnd(t.request().GetRequestID()); err != nil {
+			t.sim.logger.Error(err, "kv cache failed to process request end")
+		}
+	}
+}
+
+func (t *textCompletionReqCtx) createToolCalls() ([]openaiserverapi.ToolCall, int, string, error) {
+	return nil, 0, "", nil
+}
+
+var _ requestContext = (*textCompletionReqCtx)(nil)
+
+// Implementation of responseContext for /completions requests
+type textCompletionResponseCtx struct {
+	baseResponseContext
+}
+
+// createResponse creates the response for completion requests
+func (respCtx *textCompletionResponseCtx) createResponse() openaiserverapi.CompletionResponse {
+	baseResp := openaiserverapi.CreateBaseCompletionResponse(
+		time.Now().Unix(), respCtx.displayModelName, respCtx.usage, respCtx.id, respCtx.remoteDecode)
+	baseChoice := openaiserverapi.CreateBaseResponseChoice(0, respCtx.finishReasonPtr)
+	respText := strings.Join(respCtx.respTokens, "")
+
+	choice := openaiserverapi.CreateTextRespChoice(baseChoice, respText)
+
+	// Generate logprobs if requested for text completion
+	if respCtx.logprobs != nil && *respCtx.logprobs > 0 {
+		if logprobsData := common.GenerateTextLogprobs(respCtx.respTokens, *respCtx.logprobs); logprobsData != nil &&
+			len(logprobsData.Tokens) > 0 {
+			choice.Logprobs = logprobsData
+		} else {
+			// Set to nil if generation failed or tokens is empty
+			choice.Logprobs = nil
+		}
+	} else {
+		// Explicitly ensure logprobs is nil when not requested
+		choice.Logprobs = nil
+	}
+
+	baseResp.Object = textCompletionObject
+	return openaiserverapi.CreateTextCompletionResponse(baseResp, []openaiserverapi.TextRespChoice{choice})
+}
+
+// createUsageChunk creates and returns a CompletionRespChunk with usage data, a single chunk of streamed completion API response,
+// supports both modes
+func (respCtx *textCompletionResponseCtx) createUsageChunk() openaiserverapi.CompletionRespChunk {
+	baseChunk := openaiserverapi.CreateBaseCompletionResponse(
+		respCtx.creationTime, respCtx.displayModelName, respCtx.usageData(), respCtx.id, false)
+	baseChunk.Object = textCompletionObject
+	return openaiserverapi.CreateTextCompletionResponse(baseChunk, []openaiserverapi.TextRespChoice{})
+}
+
+// createTextCompletionChunk creates and returns a CompletionRespChunk, a single chunk of streamed completion API response,
+// for text completion.
+func (respCtx *textCompletionResponseCtx) createCompletionChunk(token string, tool *openaiserverapi.ToolCall,
+	role string, finishReason *string) openaiserverapi.CompletionRespChunk {
+	baseChunk := openaiserverapi.CreateBaseCompletionResponse(
+		respCtx.creationTime, respCtx.displayModelName, nil, respCtx.id, false)
+	baseChunk.Object = textCompletionObject
+
+	choice := openaiserverapi.CreateTextRespChoice(openaiserverapi.CreateBaseResponseChoice(0, finishReason), token)
+
+	// Generate logprobs if requested and token is not empty
+	if respCtx.logprobs != nil && token != "" && *respCtx.logprobs > 0 {
+		// Use token position based on current time
+		tokenPosition := int(respCtx.creationTime) % 1000 // Simple position simulation
+		logprobs := common.GenerateSingleTokenTextLogprobs(token, tokenPosition, *respCtx.logprobs)
+		if logprobs != nil {
+			choice.Logprobs = logprobs
+		}
+	}
+
+	return openaiserverapi.CreateTextCompletionResponse(baseChunk, []openaiserverapi.TextRespChoice{choice})
+}
+
+// in text completion there is no special first chunk
+func (respCtx *textCompletionResponseCtx) createFirstCompletionChunk() openaiserverapi.CompletionRespChunk {
+	return nil
+}
+
+func (respCtx *textCompletionResponseCtx) toolCalls() []openaiserverapi.ToolCall {
+	return nil
+}
+
+var _ responseContext = (*textCompletionResponseCtx)(nil)
